Moved Legend files to subfolder.
