+++ ts/Core/Axis/Color/ColorAxis.ts
@@ -24,7 +24,7 @@ import type ColorString from '../../Color/ColorString';
 import type ColorType from '../../Color/ColorType';
 import type Fx from '../../Animation/Fx';
 import type GradientColor from '../../Color/GradientColor';
-import type LegendOptions from '../../LegendOptions';
+import type LegendOptions from '../../Legend/LegendOptions';
 import type Point from '../../Series/Point.js';
 import type PointerEvent from '../../PointerEvent';
 import type { StatesOptionsKey } from '../../Series/StatesOptions';
@@ -38,7 +38,7 @@ import ColorAxisComposition from './ColorAxisComposition.js';
 import ColorAxisDefaults from './ColorAxisDefaults.js';
 import H from '../../Globals.js';
 const { noop } = H;
-import Legend from '../../Legend.js';
+import Legend from '../../Legend/Legend.js';
 import LegendSymbolMixin from '../../../Mixins/LegendSymbol.js';
 import SeriesRegistry from '../../Series/SeriesRegistry.js';
 const { series: Series } = SeriesRegistry;
