+++ ts/Core/Legend/LegendLike.d.ts
